+++ R/globals.R
@@ -8,6 +8,14 @@ utils::globalVariables(c(
   "n_projected", # <aggregate_measures>
   "n_measure", # <aggregate_measures>
   "wmape", # <aggregate_measures>
+  "age", # <calculate_shares>
+  "sum_5", # <calculate_shares>
+  "sum_10", # <calculate_shares>
+  "prop_5", # <calculate_shares>
+  "prop_10", # <calculate_shares>
+  "nat", # <calculate_shares>
+  "sex", # <calculate_shares>
+  "n_sum", # <calculate_shares>
   "mig_sub", # <complement_projection>
   "age", # <complement_projection>
   "n_jan", # <complement_projection>